correction of solrmarc vs. solrdefault functions. Fixes some thumbs
