+++ module/VufindUNL/config/module.config.php
@@ -10,6 +10,7 @@ return array (
         'factories' => 
         array (
           'solrmarc' => 'VufindUNL\\RecordDriver\\Factory::getSolrMarc',
+          'solrdefault' => 'VufindUNL\\RecordDriver\\Factory::getSolrDefault',
         ),
       ),
     ),
